+++ lib/acts_as_list.rb
@@ -4,5 +4,5 @@ require "acts_as_list/active_record/acts/top_of_list_method_definer"
 require "acts_as_list/active_record/acts/add_new_at_method_definer"
 require "acts_as_list/active_record/acts/aux_method_definer"
 require "acts_as_list/active_record/acts/callback_definer"
-require "acts_as_list/active_record/acts/shuffle_positions_on_intermediate_items__method_definer"
+require "acts_as_list/active_record/acts/shuffle_positions_on_intermediate_items_method_definer"
 require 'acts_as_list/active_record/acts/list'
