+++ lib/acts_as_list/active_record/acts/shuffle_positions_on_intermediate_items_method_definer.rb
